% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCooccur_functions.R
\name{calc_co_occurrence_for_radius}
\alias{calc_co_occurrence_for_radius}
\title{Calculate Co-occurrence Matrix for a Given Radius}
\usage{
calc_co_occurrence_for_radius(
  seurat_obj,
  radius,
  sample_key,
  cluster_key,
  k = 30
)
}
\arguments{
\item{seurat_obj}{Seurat object with spatial coordinates.}

\item{radius}{Radius to define local neighborhood.}

\item{sample_key}{Metadata column specifying sample identity.}

\item{cluster_key}{Metadata column specifying cluster labels.}

\item{k}{Maximum number of neighbors to consider.}
}
\value{
A list with co-occurrence count and enrichment ratio matrices.
}
\description{
Calculate Co-occurrence Matrix for a Given Radius
}
