% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCooccur_functions.R
\name{compute_co_occurrence_ratio}
\alias{compute_co_occurrence_ratio}
\title{Compute Enrichment Ratios from Count Matrix}
\usage{
compute_co_occurrence_ratio(co_occur_count)
}
\arguments{
\item{co_occur_count}{Matrix of observed co-occurrence counts.}
}
\value{
A matrix of normalized enrichment ratios.
}
\description{
Compute Enrichment Ratios from Count Matrix
}
