% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCooccur_functions.R
\name{compute_count}
\alias{compute_count}
\title{Compute Co-occurrence Count Matrix}
\usage{
compute_count(
  adj,
  int_clust_row,
  int_clust_col,
  n_cls,
  cluster_data,
  transformation = TRUE
)
}
\arguments{
\item{adj}{Adjacency matrix.}

\item{int_clust_row}{Vector of cluster labels for rows.}

\item{int_clust_col}{Vector of cluster labels for columns.}

\item{n_cls}{Number of clusters.}

\item{cluster_data}{Original cluster assignments.}

\item{transformation}{Whether to transform counts based on adjacency normalization.}
}
\value{
A co-occurrence count matrix.
}
\description{
Compute Co-occurrence Count Matrix
}
