% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCooccur_functions.R
\name{cooccur_local}
\alias{cooccur_local}
\title{Local Co-occurrence Score (Generic method)}
\usage{
cooccur_local(
  df,
  cluster_x,
  cluster_y,
  connectivity_key = "nn",
  neighbors.k = 20,
  radius = 30,
  maxnsteps = 1
)
}
\arguments{
\item{df}{Data.frame of coordinates and cluster.}

\item{cluster_x}{First cluster of interest.}

\item{cluster_y}{Second cluster of interest.}

\item{connectivity_key}{Graph type to use.}

\item{neighbors.k}{Number of neighbors.}

\item{radius}{Radius for neighborhood.}

\item{maxnsteps}{Number of diffusion steps.}
}
\value{
Data.frame with scores.
}
\description{
Local Co-occurrence Score (Generic method)
}
