% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCooccur_functions.R
\name{cooccur_local.Seurat}
\alias{cooccur_local.Seurat}
\title{Local Co-occurrence Score (Seurat Method)}
\usage{
cooccur_local.Seurat(
  seurat_obj,
  cluster_x,
  cluster_y,
  connectivity_key = "nn",
  cluster_key = "seurat_clusters",
  sample_key = "sample_id",
  neighbors.k = 20,
  radius = 30,
  maxnsteps = 15
)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{cluster_x}{First cluster of interest.}

\item{cluster_y}{Second cluster of interest.}

\item{connectivity_key}{Graph type to use.}

\item{cluster_key}{Metadata column with cluster info.}

\item{sample_key}{Metadata column with sample ID.}

\item{neighbors.k}{Number of neighbors to build graph.}

\item{radius}{Radius for proximity-based interaction.}

\item{maxnsteps}{Maximum diffusion steps.}
}
\value{
A data.frame with local co-occurrence scores.
}
\description{
Local Co-occurrence Score (Seurat Method)
}
