% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCooccur_functions.R
\docType{data}
\name{manual_colors}
\alias{manual_colors}
\title{Default Manual Colors for Clusters}
\format{
An object of class \code{character} of length 31.
}
\usage{
manual_colors
}
\description{
Default Manual Colors for Clusters
}
\keyword{datasets}
