% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCooccur_functions.R
\name{nhood_enrichment}
\alias{nhood_enrichment}
\title{Neighborhood Enrichment (Generic method)}
\usage{
nhood_enrichment(
  df,
  cluster_key,
  neighbors.k = 30,
  connectivity_key = "nn",
  transformation = TRUE,
  n_perms = 100,
  seed = 1938493,
  n_jobs = 4
)
}
\arguments{
\item{df}{Data.frame with spatial and cluster metadata.}

\item{cluster_key}{Column with cluster labels.}

\item{neighbors.k}{Number of neighbors to use.}

\item{connectivity_key}{Type of graph: "nn" or "snn".}

\item{transformation}{Whether to normalize adjacency matrix.}

\item{n_perms}{Number of permutations.}

\item{seed}{Random seed.}

\item{n_jobs}{Number of parallel jobs.}
}
\value{
A list with z-score and count matrices.
}
\description{
Neighborhood Enrichment (Generic method)
}
