% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCooccur_functions.R
\name{nhood_enrichment.Seurat}
\alias{nhood_enrichment.Seurat}
\title{Neighborhood Enrichment (Seurat Method)}
\usage{
nhood_enrichment.Seurat(
  seurat_obj,
  cluster_key,
  neighbors.k = 30,
  connectivity_key = "nn",
  transformation = TRUE,
  n_perms = 100,
  seed = 1938493,
  n_jobs = 4
)
}
\arguments{
\item{seurat_obj}{A Seurat object with spatial coordinates.}

\item{cluster_key}{Metadata column for cluster IDs.}

\item{neighbors.k}{Number of neighbors to construct graph.}

\item{connectivity_key}{Which graph to use: "nn" or "snn".}

\item{transformation}{Logical, whether to normalize adjacency matrix.}

\item{n_perms}{Number of permutations for significance testing.}

\item{seed}{Random seed for reproducibility.}

\item{n_jobs}{Number of cores to use in parallel.}
}
\value{
Updated Seurat object with z-scores in misc slot.
}
\description{
Neighborhood Enrichment (Seurat Method)
}
