% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCooccur_functions.R
\name{permute_clusters}
\alias{permute_clusters}
\title{Permute Cluster Assignments and Recompute Counts}
\usage{
permute_clusters(adj, int_clust, n_cls, cluster_data, transformation)
}
\arguments{
\item{adj}{Adjacency matrix.}

\item{int_clust}{Cluster labels.}

\item{n_cls}{Number of clusters.}

\item{cluster_data}{Original cluster data.}

\item{transformation}{Whether to apply adjacency transformation.}
}
\value{
Permuted co-occurrence count matrix.
}
\description{
Permute Cluster Assignments and Recompute Counts
}
