% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_compare_plots.R
\name{plot_volcano_groups}
\alias{plot_volcano_groups}
\title{Volcano plot of a compare_groups result}
\usage{
plot_volcano_groups(
  compare_df,
  effect_col = "effect",
  p_col = "padj",
  sig_threshold = 0.05,
  label_top = 10,
  pair_keys = c("cluster_i", "cluster_j")
)
}
\arguments{
\item{compare_df}{Output of [compare_groups()].}

\item{effect_col}{Column for the x axis. Defaults to "effect".}

\item{p_col}{Column for the y axis (will be -log10 transformed).
Defaults to "padj".}

\item{sig_threshold}{Adjusted p-value threshold for highlighting.}

\item{label_top}{Integer, number of top pairs to label by p-value.
Set to 0 to skip labels.}

\item{pair_keys}{Columns identifying the cluster pair (used for labels).}
}
\value{
A ggplot object.
}
\description{
Volcano plot of a compare_groups result
}
