% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCooccur_functions.R
\name{search_interaction_spot}
\alias{search_interaction_spot}
\title{Search for Spatial Interaction Spots}
\usage{
search_interaction_spot(
  seurat_object,
  fov,
  radius,
  n_min,
  neighbors.k = 200,
  cell_id = cell_id,
  cluster_col = cluster_col,
  target_cluster = target_cluster
)
}
\arguments{
\item{seurat_object}{Seurat object.}

\item{fov}{Field of view identifier.}

\item{radius}{Radius threshold for neighborhood.}

\item{n_min}{Minimum number of cells to qualify as a spot.}

\item{neighbors.k}{Max number of neighbors to consider.}

\item{cell_id}{Vector of target cell IDs.}

\item{cluster_col}{Column name in metadata specifying cluster assignment.}

\item{target_cluster}{Target cluster(s) assumed to be interacting.}
}
\value{
A data.frame of detected interaction clusters and metadata.
}
\description{
Search for Spatial Interaction Spots
}
